Added even distribution
Now we can add ('distribution': 'even') to any entity to set equal
chances for its sentences to be selected when generating examples.

Sometimes when we have a long list of sentences which we want to have
the same chance of appearing in the examples, it is very tedious
to set the probabilities.
It is even impossible if we have let's say three sentences because
the probability operator allows only integer values.